@@ -1,4 +1,5 @@
 import 'package:example/mock_data.dart';
+import 'package:example/widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:sticky_and_expandable_list/sticky_and_expandable_list.dart';
 
@@ -25,7 +26,7 @@ class _ExampleAnimatableHeaderState extends State<ExampleAnimatableHeader> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: Text("Animatable Header Example")),
+      appBar: AppBar(title: TitleText("Animatable Header Example")),
       body: CustomScrollView(
         slivers: <Widget>[
           SliverToBoxAdapter(
@@ -50,7 +51,7 @@ class _ExampleAnimatableHeaderState extends State<ExampleAnimatableHeader> {
                     leading: CircleAvatar(
                       child: Text("$index"),
                     ),
-                    title: Text(item),
+                    title: TitleText(item),
                   );
                 }),
           )
@@ -67,7 +68,7 @@ class _ExampleAnimatableHeaderState extends State<ExampleAnimatableHeader> {
         double opacity = _controller.switchingSectionIndex == sectionIndex
             ? (1 - _controller.percent)
             : 1;
-        String headerText = section.header;
+        String headerText = section.header ?? "";
         if (_controller.switchingSectionIndex == sectionIndex) {
           headerText += " Switching";
         } else if (_controller.stickySectionIndex == sectionIndex) {
