@@ -1,4 +1,5 @@
 import 'package:example/mock_data.dart';
+import 'package:example/widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:sticky_and_expandable_list/sticky_and_expandable_list.dart';
 
@@ -21,7 +22,7 @@ class _ExampleListViewState extends State<ExampleListView> {
     //class ExampleSection implements ExpandableListSection<String> {}
     //so: SliverExpandableChildDelegate<String, ExampleSection>()
     return Scaffold(
-        appBar: AppBar(title: Text("ListView Example")),
+        appBar: AppBar(title: TitleText("ListView Example")),
         body: ExpandableListView(
           builder: SliverExpandableChildDelegate<String, ExampleSection>(
               sectionList: sectionList,
