@@ -1,4 +1,5 @@
 import 'package:example/mock_data.dart';
+import 'package:example/widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:sticky_and_expandable_list/sticky_and_expandable_list.dart';
 
@@ -26,7 +27,7 @@ class _ExampleSideHeaderState extends State<ExampleSideHeader> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(title: Text("Side Header Example")),
+        appBar: AppBar(title: TitleText("Side Header Example")),
         body: ExpandableListView(
           builder: SliverExpandableChildDelegate<String, ExampleSection>(
               overlapsContent: true,
